requirements.txt
django~=3.1.0
djangorestframework~=3.12.0
django-rest-auth~=0.9
django-allauth==0.42.0
djangorestframework-jwt~=1.11.0
django-cors-headers~=3.6.0
